Generated image file for how3ctf
Once the event file is made, just run the ./event-images.py script and
it will pop out the necessary image. You need to specify the path for
the image manually in the event file.

For this to work, you'll probably need to install `python-pil`, PIL or
Pillow or something similar from your package manager or pip.